clean code and package rename
