--- oam-network-agent/src/main/java/org/xgvela/config/HeartBeatConfiguration.java
@@ -1,13 +1,0 @@
-package org.xgvela.config;
-
-import lombok.Data;
-import org.springframework.boot.context.properties.ConfigurationProperties;
-import org.springframework.stereotype.Component;
-
-@Data
-@Component
-@ConfigurationProperties(prefix = "heartbeat")
-public class HeartBeatConfiguration {
-    private String host;
-    private Integer port;
-}
