--- oam-network-agent/src/main/java/org/xgvela/entity/UpdateFileCallBack.java
@@ -1,15 +1,0 @@
-package org.xgvela.entity;
-
-public class UpdateFileCallBack {
-
-    private String result;
-
-    public String getResult() {
-        return result;
-    }
-
-    public void setResult(String result) {
-        this.result = result;
-    }
-
-}
