+++ oam-network-agent/src/main/java/org/xgvela/oam/controller/IMetricController.java
@@ -1,0 +1,17 @@
+package org.xgvela.oam.controller;
+
+import io.swagger.annotations.Api;
+import io.swagger.annotations.ApiOperation;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+@Api(tags = "Proxy layer - performance ")
+@RestController
+@RequestMapping(value = "/metricTest")
+public interface IMetricController {
+
+    @ApiOperation(value = "metric performance data ", notes = "metric performance data ")
+    @GetMapping
+    String MetricData();
+}
