+++ oam-network-agent/src/main/java/org/xgvela/oam/entity/UpdateFileCallBack.java
@@ -1,0 +1,15 @@
+package org.xgvela.oam.entity;
+
+public class UpdateFileCallBack {
+
+    private String result;
+
+    public String getResult() {
+        return result;
+    }
+
+    public void setResult(String result) {
+        this.result = result;
+    }
+
+}
