+++ oam-network-agent/src/main/java/org/xgvela/oam/service/performance/IMeterService.java
@@ -1,0 +1,9 @@
+package org.xgvela.oam.service.performance;
+
+public interface IMeterService {
+    void MetricData();
+
+    void MetricGrpcValue(String neId, String neType, String metric, Double value);
+
+    void MetricGrpcDefaultValue(String neId, String neType);
+}
