+++ oam-network-agent/src/main/java/org/xgvela/oam/service/signal/SignalingTraceService.java
@@ -1,0 +1,10 @@
+package org.xgvela.oam.service.signal;
+
+import org.springframework.stereotype.Component;
+
+/**
+ *
+ */
+@Component
+public class SignalingTraceService {
+}
