--- oam-network-agent/src/main/java/org/xgvela/service/heartbeat/IOamVnfSelectService.java
@@ -1,7 +1,0 @@
-package org.xgvela.service.heartbeat;
-
-import com.baomidou.mybatisplus.extension.service.IService;
-import org.xgvela.oam.entity.nftube.OamVnf;
-
-public interface IOamVnfSelectService extends IService<OamVnf> {
-}
