--- oam-network-agent/src/main/java/org/xgvela/service/performance/IMeterService.java
@@ -1,9 +1,0 @@
-package org.xgvela.service.performance;
-
-public interface IMeterService {
-    void MetricData();
-
-    void MetricGrpcValue(String neId, String neType, String metric, Double value);
-
-    void MetricGrpcDefaultValue(String neId, String neType);
-}
