--- oam-network-agent/src/main/java/org/xgvela/service/signal/SignalingTraceService.java
@@ -1,10 +1,0 @@
-package org.xgvela.service.signal;
-
-import org.springframework.stereotype.Component;
-
-/**
- *
- */
-@Component
-public class SignalingTraceService {
-}
