--- oam-network-agent/src/main/java/org/xgvela/utils/SpringUtils.java
@@ -1,36 +1,0 @@
-package org.xgvela.utils;
-
-import org.springframework.context.ApplicationContext;
-import org.springframework.context.ApplicationContextAware;
-import org.springframework.core.env.Environment;
-import org.springframework.stereotype.Component;
-
-@Component
-public class SpringUtils implements ApplicationContextAware {
-    private static ApplicationContext applicationContext;
-
-    @Override
-    public void setApplicationContext(ApplicationContext applicationContext) {
-        SpringUtils.applicationContext = applicationContext;
-    }
-
-    public static ApplicationContext getContext() {
-        return applicationContext;
-    }
-
-    public static <T> T getBean(Class<T> clazz) {
-        return applicationContext.getBean(clazz);
-    }
-
-    public static <T> T getBean(String name, Class<T> clazz) {
-        return applicationContext.getBean(name, clazz);
-    }
-
-    public static Object getBean(String name) {
-        return applicationContext.getBean(name);
-    }
-
-    public static Environment getEnv() {
-        return applicationContext.getEnvironment();
-    }
-}
